+++ src/schemathesis/cli/options.py
@@ -1,5 +1,5 @@
 from enum import Enum
-from typing import List, Optional, Type
+from typing import List, Optional, Type, Union
 
 import click
 
@@ -20,3 +20,22 @@ class CSVOption(click.Choice):
         sorted_options = ", ".join(sorted(invalid_options, key=items.index))
         available_options = ", ".join(self.choices)
         self.fail(f"invalid choice(s): {sorted_options}. Choose from {available_options}")
+
+
+class NotSet:
+    pass
+
+
+not_set = NotSet()
+
+
+class OptionalInt(click.types.IntParamType):
+    def convert(  # type: ignore
+        self, value: str, param: Optional[click.core.Parameter], ctx: Optional[click.core.Context]
+    ) -> Union[int, NotSet]:
+        if value == "None":
+            return not_set
+        try:
+            return int(value)
+        except (ValueError, UnicodeError):
+            self.fail("%s is not a valid integer or None" % value, param, ctx)